Move MergeInfo out of InitializedStateDetails
